place controller-type-extractor next to resolvers
https://coveord.atlassian.net/browse/KIT-682
